requirements_local.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
Pillow==8.3.1
pkg_resources==0.0.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
